requirements.txt
azure-functions==1.6.0
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.20.3
pandas==1.2.4
pyodbc==4.0.30
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
urllib3==1.26.4
